Fixup interval joiners Javadoc, move impl. to Joiners, add missing test cases
